Twitter part 2

+++ controllers/users.controller.js
@@ -1,23 +1,19 @@
 import { db } from "../config/db.js";
-import { ObjectId } from "mongodb";
 import { userRequest } from "../validators/user.schema.js";
 import { createUser as createUserService } from "../services/createUser.service.js";
 
 export async function createUser(req, res) {
     try {
-        // const authHeader = req.headers.authorization;
-        // if (!authHeader?.startsWith("Bearer ")) {
-        //     return res.status(401).send("Unauthorized");
-        // }
-        // const token = authHeader.split(" ")[1];
-        // const oauthUser = await verifyToken(token);
+        if (!req.isAuthenticated()) {
+            return res.status(401).send("Unauthorized: Please log in");
+        }
 
         const { error } = userRequest.validate(req.body);
         if (error) {
             return res.status(400).send(error.details[0].message);
         }
-        await createUserService(req.body, "oauthplaceholder123");
-        res.status(201).send("User oauthplaceholder123 created");
+        await createUserService(req.body, req.user.id);
+        res.status(201).send(`User ${req.user.id} created`);
     } catch (err) {
         // https://www.mongodb.com/docs/manual/reference/error-codes/ says 11000 is duplicate key error
         if (err.code === 11000) {
@@ -28,8 +24,16 @@ export async function createUser(req, res) {
 }
 
 export async function updateUser(req, res) {
-    const id = req.params.id;
     try {
+        if (!req.isAuthenticated()) {
+            return res.status(401).send("Unauthorized: Please log in");
+        }
+
+        const id = req.params.id;
+        if (id !== req.user.id) {
+            return res.status(403).send("Forbidden: You can only modify your own resources");
+        }
+
         const { error } = userRequest.validate(req.body);
         if (error) {
             return res.status(400).send(error.details[0].message);
@@ -60,8 +64,16 @@ export async function getUserById(req, res) {
 }
 
 export async function deleteUser(req, res) {
-    const id = req.params.id;
     try {
+        if (!req.isAuthenticated()) {
+            return res.status(401).send("Unauthorized: Please log in");
+        }
+
+        const id = req.params.id;
+        if (id !== req.user.id) {
+            return res.status(403).send("Forbidden: You can only modify your own resources");
+        }
+
         const result = await db.collection("users").deleteOne({ _id: id });
 
         if (result.deletedCount === 0) {
